+++ butano/include/btn_hash_map_fwd.h
@@ -8,8 +8,7 @@ namespace btn
     template<typename Key, typename Value, typename KeyHash = hash<Key>, typename KeyEqual = equal_to<Key>>
     class ihash_map;
 
-    template<typename Key, typename Value, size_t MaxSize, typename KeyHash = hash<Key>,
-             typename KeyEqual = equal_to<Key>>
+    template<typename Key, typename Value, int MaxSize, typename KeyHash = hash<Key>, typename KeyEqual = equal_to<Key>>
     class hash_map;
 }
